Decoder works
Needed to add state to track whether the header has been parsed or not

+++ src/cast/payload/mod.rs
@@ -8,15 +8,15 @@ pub enum Payload {
     Ping,
     Pong,
     #[serde(rename_all = "camelCase")]
-    Launch { request_id: i64, app_id: String },
+    Launch { request_id: i32, app_id: String },
     #[serde(rename_all = "camelCase")]
-    Stop { request_id: i64, session_id: String },
+    Stop { request_id: i32, session_id: String },
     #[serde(rename_all = "camelCase")]
-    GetStatus { request_id: i64 },
+    GetStatus { request_id: i32 },
     #[serde(rename_all = "camelCase")]
-    GetAppAvailability { request_id: i64, app_id: Vec<String> },
+    GetAppAvailability { request_id: i32, app_id: Vec<String> },
     #[serde(rename_all = "camelCase")]
-    ReceiverStatus { request_id: i64, status: Status },
+    ReceiverStatus { request_id: i32, status: Status },
     SetVolume { volume: Volume },
 }
